pyproject.toml
[tool.poetry]
name = "test-for-onescreendi"
version = "0.1.0"
description = ""
authors = ["mr360 <vadgr1schin@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.7.4"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.31"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
